requirements.txt
discord.py==1.0.0
git+https://gitlab.com/neko404notfound/libneko
Pillow
aiohttp>=3.3.0,<3.4.0
websockets<7.0,>=6.0
aiofiles
psutil
jishaku